Feat: search restaurant distance type to integer

+++ src/main/java/com/zerobase/hseungho/restaurantreservation/service/dto/restaurant/IRestaurantDto.java
@@ -11,7 +11,7 @@ public interface IRestaurantDto {
     Long getId();
     String getName();
     String getAddress();
-    Double getDistance();
+    Integer getDistance();
     String getDescription();
     Integer getOpenHour();
     Integer getOpenMinute();
